requirements.txt
anytree==2.8.0
certifi==2022.12.7
charset-normalizer==2.1.1
diskcache==5.4.0
idna==3.4
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.14
xkcd2347==0.0.4
